acces à la page de création d'équipe à partir de l'index.
